adding some fileds to Organizer's Model, structuring API file organization and more
