+++ EventBliss2.0/package.json
@@ -26,7 +26,7 @@
     "react-image-upload": "^3.0.1",
     "react-parallax": "^3.5.1",
     "react-router-dom": "^6.22.3",
-    "sweetalert2": "^11.10.7",
+    "sweetalert2": "^11.6.13",
     "swiper": "^11.0.7"
   },
   "devDependencies": {
