+++ EventBliss2.0/src/components/api/event/index.jsx
@@ -1,0 +1,6 @@
+import axios  from 'axios'
+
+//----------EVENTS----------//
+export const eventApi = axios.create({
+    baseURL: 'http://localhost:8000/api/v1/events/',
+});
