+++ EventBliss2.0/src/components/api/organizer/index.jsx
@@ -1,0 +1,12 @@
+import axios from 'axios'
+
+//----------ORGANIZER----------//
+export const organizerAPI = axios.create({
+    baseURL: 'http://localhost:8000/api/v1/organizers/',
+});
+  
+
+
+
+  
+
