+++ EventBliss2.0/src/components/api/organizer/post.jsx
@@ -1,0 +1,6 @@
+import { organizerAPI } from ".";
+
+export const createOrganizer = (organizer) => {
+    return organizerAPI.post('/',organizer);
+};
+
