+++ EventBliss2.0/src/components/api/request/index.jsx
@@ -1,0 +1,9 @@
+import axios from 'axios'
+
+//----------REQUEST----------//
+export const eventRequestsAPI = axios.create({
+    baseURL: 'http://localhost:8000/api/v1/eventRequest/',
+  });
+  
+
+
