+++ EventBliss2.0/src/components/api/request/put.jsx
@@ -1,0 +1,11 @@
+import { eventRequestsAPI } from "."
+
+/**
+ * This funtion update the data of an event request
+ * @param {*} id request id
+ * @param {*} the new data (object)
+ * @returns 
+ */
+export const updateEventRequest = (id,data) => {
+    return eventRequestsAPI.put(`/${id}/`,data)
+}
