+++ EventBliss2.0/src/components/user.jsx
@@ -1,5 +1,4 @@
-import { DataBackend } from "../DataBackend";
-
+import { DataBackend } from "./api/DataBackend";
 export function User() {
     DataBackend();
     return (
