+++ EventBliss2.0/src/main.jsx
@@ -5,7 +5,7 @@ import './index.css'
 import { ClerkProvider } from '@clerk/clerk-react'
 
 
-const PUBLISHABLE_KEY = import.meta.env.VITE_CLERK_PUBLISHABLE_KEY
+const PUBLISHABLE_KEY = "pk_test_dG91Z2gtcmFjY29vbi00Mi5jbGVyay5hY2NvdW50cy5kZXYk"
 if (!PUBLISHABLE_KEY) {
   throw new Error("Missing Publishable Key")
 }
