+++ client/serializer.py
@@ -1,6 +1,6 @@
 from rest_framework import serializers
 
-from .models import Client,Organizer
+from .models import Category, Client,Organizer
 
 class ClientSerializer(serializers.ModelSerializer):
     class Meta:
@@ -11,3 +11,9 @@ class OrganizerSerializer(serializers.ModelSerializer):
     class Meta:
         model = Organizer
         fields = '__all__'
+
+
+class CategorySerializer(serializers.ModelSerializer):
+    class Meta:
+        model = Category
+        fields= '__all__'
