+++ client/views.py
@@ -1,6 +1,6 @@
-from .models import Client,Organizer
+from .models import Client,Organizer,Category
 from rest_framework import viewsets
-from .serializer import ClientSerializer,OrganizerSerializer
+from .serializer import ClientSerializer,OrganizerSerializer,CategorySerializer
 # Create your views here.
 
 
@@ -12,3 +12,8 @@ class ClientView(viewsets.ModelViewSet):
 class OrganizerView(viewsets.ModelViewSet):
     serializer_class = OrganizerSerializer
     queryset = Organizer.objects.all()
+
+
+class CategoryView(viewsets.ModelViewSet):
+    serializer_class = CategorySerializer
+    queryset = Category.objects.all()
